+++ requirements.txt
@@ -1,0 +1,3 @@
+pytest>=7.0.0
+pytest-subtests
+pyyaml
