--- requiremetnts.txt
@@ -1,2 +1,0 @@
-pytest>=7.0.1
-pyyaml
