+++ setup.cfg
@@ -17,3 +17,6 @@ exclude =
     old,
     build,
     dist
+
+[mypy]
+ignore_missing_imports = True
